requirements.txt
requests
gspread
google-auth
python-dotenv
Flask